refactor: update auth layout structure
