+++ src/app/[locale]/(dashboard)/layout.tsx
@@ -16,9 +16,11 @@ export default async function DashboardLayout({
   }
   return (
     <>
+      <Header />
       <main className="main flex flex-col justify-center items-center gap-40 w-full max-w-[144rem] my-0 mx-auto px-16 py-0">
         {children}
       </main>
+      <Footer />
     </>
   );
 }
